requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
SQLAlchemy==2.0.7
Werkzeug==2.2.3
python-dotenv==1.0.0
psycopg2-binary==2.9.6
openai==1.3.0
stripe==2.60.0
python-jose==3.3.0
passlib==1.7.4
Pillow==9.5.0
python-multipart==0.0.6
requests==2.28.2 